requirements.txt
flask~=1.1.2
gunicorn
psycopg2-binary
flask-cors